+++ packages/components/icons/scripts/utils/tagify.mjs
@@ -2,9 +2,7 @@ import { camelCase } from 'change-case'
 
 const tagify = componentsMap =>
   `${Array.from(componentsMap.entries())
-    .map(
-      ([name, { dir }]) => `export { tags as ${camelCase(`${name}Tags`)} } from './icons/${name}'`
-    )
+    .map(([name]) => `export { tags as ${camelCase(`${name}Tags`)} } from './icons/${name}'`)
     .join('\n')}\n`
 
 export default tagify